removed unused tests code. (#4093)
